requirements.txt
tkinter
Pillow>=9.0.0
pygame>=2.0.0
pyinstaller>=5.0.0